requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
_pytorch_select=0.2=gpu_0
attrs=19.3.0=py_0
backcall=0.1.0=py_0
blas=1.0=mkl
bleach=3.1.5=pyh9f0ad1d_0
brotlipy=0.7.0=py38h1e0a361_1000
bzip2=1.0.8=h516909a_2
ca-certificates=2020.6.20=hecda079_0
cairo=1.16.0=hcf35c78_1003
certifi=2020.6.20=py38h32f6830_0
cffi=1.14.0=py38hd463f26_0
chardet=3.0.4=py38h32f6830_1006
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.3.0=py_0
cryptography=2.9.2=py38h766eaa4_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py_2
dbus=1.13.6=he372182_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py38h32f6830_1001
expat=2.2.9=he1b5a44_2
ffmpeg=4.2=h167e202_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.1=he06d7ca_0
fribidi=1.0.9=h516909a_0
future=0.18.2=py38h32f6830_1
gettext=0.19.8.1=hc5be6a0_1002
glib=2.64.2=h6f030ca_0
gmp=6.2.0=he1b5a44_2
gnutls=3.6.5=hd3a4fd2_1002
graphite2=1.3.13=he1b5a44_1001
graphviz=2.42.3=h0511662_0
grpcio=1.28.1=pypi_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
gym=0.17.1=py38h32f6830_0
gym-unity=0.16.0=pypi_0
harfbuzz=2.4.0=h9f30f68_3
icu=64.2=he1b5a44_1
idna=2.9=py_1
importlib-metadata=1.6.0=py38h32f6830_0
importlib_metadata=1.6.0=0
ipykernel=5.2.1=py38h23f93f0_0
ipython=7.14.0=py38h32f6830_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.5.1=py_0
jedi=0.17.0=py38h32f6830_0
jinja2=2.11.2=pyh9f0ad1d_0
jpeg=9c=h14c3975_1001
json5=0.9.0=py_0
jsonschema=3.2.0=py38h32f6830_1
jupyter=1.0.0=py_2
jupyter_client=6.1.3=py_0
jupyter_console=6.1.0=py_1
jupyter_core=4.6.3=py38h32f6830_1
jupyterlab=2.1.2=py_0
jupyterlab_server=1.1.1=py_0
kiwisolver=1.2.0=py38hbf85e49_0
lame=3.100=h14c3975_1001
ld_impl_linux-64=2.34=h53a641e_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libclang=9.0.1=default_hde54327_0
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.2.0=h24d8f2e_2
libgfortran-ng=7.3.0=hdf63c60_5
libiconv=1.15=h516909a_1006
liblapack=3.8.0=14_mkl
libllvm9=9.0.1=he513fc3_1
libpng=1.6.37=hed695b0_1
libsodium=1.0.17=h516909a_0
libstdcxx-ng=9.2.0=hdf63c60_2
libtiff=4.1.0=hc7e4089_6
libtool=2.4.6=h14c3975_1002
libuuid=2.32.1=h14c3975_1000
libuv=1.34.0=h516909a_0
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
llvm-openmp=10.0.0=hc9558a2_0
lz4-c=1.9.2=he1b5a44_1
markupsafe=1.1.1=py38h1e0a361_1
matplotlib=3.2.1=0
matplotlib-base=3.2.1=py38h2af1d28_0
mistune=0.8.4=py38h1e0a361_1001
mkl=2019.5=281
mkl-service=2.3.0=py38h516909a_0
mlagents-envs=0.16.0=pypi_0
mpi=1.0=openmpi
mpi4py=3.0.3=py38h246a051_1
nbconvert=5.6.1=py38h32f6830_1
nbformat=5.0.6=py_0
ncurses=6.1=hf484d3e_1002
neat-python=0.92=pypi_0
nettle=3.4.1=h1bed415_1002
ninja=1.10.0=hc9558a2_0
nodejs=13.13.0=hf5d1a2b_0
notebook=6.0.3=py38_0
nspr=4.25=he1b5a44_0
nss=3.47=he751ad9_0
numpy=1.18.4=py38h8854b6b_0
openh264=1.8.0=hdbcaa40_1000
openmpi=4.0.4=hdf1f1ad_0
openssl=1.1.1g=h516909a_0
packaging=20.1=py_0
pandoc=2.9.2.1=0
pandocfilters=1.4.2=py_1
pango=1.42.4=h7062337_4
parso=0.7.0=pyh9f0ad1d_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=py38h32f6830_1
pickleshare=0.7.5=py38h32f6830_1001
pillow=7.1.2=pypi_0
pip=20.1=pyh9f0ad1d_0
pixman=0.38.0=h516909a_1003
prometheus_client=0.7.1=py_0
prompt-toolkit=3.0.5=py_0
prompt_toolkit=3.0.5=0
protobuf=3.11.3=pypi_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py_1001
pybullet=2.8.3=pypi_0
pycparser=2.20=py_0
pyglet=1.5.0=py38h32f6830_1
pygments=2.6.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py38ha8c2ead_3
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.16.0=py38h1e0a361_0
pysocks=1.7.1=py38h32f6830_1
python=3.8.2=he5300dc_7_cpython
python-dateutil=2.8.1=py_0
python-graphviz=0.14=pyh9f0ad1d_0
python_abi=3.8=1_cp38
pytorch=1.4.0=cuda101py38h02f0884_0
pyzmq=19.0.0=py38ha71036d_1
qt=5.12.5=hd8c4c69_1
qtconsole=4.7.3=pyh9f0ad1d_0
qtpy=1.9.0=py_0
readline=8.0=hf8c457e_0
requests=2.23.0=pyh8c360ce_2
scipy=1.4.1=py38h18bccfc_3
send2trash=1.5.0=py_0
setuptools=46.1.3=py38h32f6830_0
six=1.14.0=py_1
sqlite=3.30.1=hcee41ef_0
style=1.1.0=pypi_0
terminado=0.8.3=py38h32f6830_1
testpath=0.4.4=py_0
tk=8.6.10=hed695b0_0
tornado=6.0.4=py38h1e0a361_1
traitlets=4.3.3=py38h32f6830_1
urllib3=1.25.9=py_0
wcwidth=0.1.9=pyh9f0ad1d_0
webencodings=0.5.1=py_1
wheel=0.34.2=py_1
widgetsnbextension=3.5.1=py38_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxpm=3.5.13=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-libxt=1.1.5=h516909a_1003
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h516909a_0
zeromq=4.3.2=he1b5a44_2
zipp=3.1.0=py_0
zlib=1.2.11=h516909a_1006
zstd=1.4.4=h6597ccf_3
